Some sign mismatches fixed.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@482 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CXMLReaderImpl.h
@@ -506,7 +506,7 @@ private:
 
 		core::string<char_type> newstr;
 
-		while(pos != -1 && pos < origstr.size()-2)
+		while(pos != -1 && pos < (int)origstr.size()-2)
 		{
 			// check if it is one of the special characters
 
@@ -539,7 +539,7 @@ private:
 			pos = origstr.findNext(L'&', pos);		
 		}
 
-		if (oldPos < origstr.size()-1)
+		if (oldPos < (int)origstr.size()-1)
 			newstr.append(origstr.subString(oldPos, origstr.size()-oldPos));
 
 		return newstr;
